+++ src/YTTrending.API/GlobalUsings.cs
@@ -3,7 +3,7 @@ global using Microsoft.AspNetCore.Mvc;
 
 global using YTTrending.Infrastructure;
 global using YTTrending.Infrastructure.Persistence;
-
-// Thêm khi bước 5/6 tạo ra chúng:
+global using YTTrending.Application;   // AddApplication()
 // global using YTTrending.Application.Common;
-// global using YTTrending.API.Common;
+global using YTTrending.API.Common;
+global using YTTrending.Application.Common.Models;